rust: fix self-referencing symlink by targeting parent directory instead
Prior to this commit the symlink was placed at `{dir}/{src_unit}`,
pointing to `{src_unit}`, thus referencing itself. Now it points to the
parent directory instead, where the actual service file is located.
I didn't further investigate to why this wasn't an issue prior to
systemd v258, but this fixes the generator for the release.

I did not opt for a generic approach here because I couldn't see a need
for it.